config image upload

+++ database/migrations/2022_08_28_012510_create_periods_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name')->index();
             $table->unsignedBigInteger('term')->nullable();
             $table->foreign('term')->references('id')->on('terms');
-            $table->tinyInteger('year')->index();
+            $table->integer('year')->index();
             $table->date('class_start')->nullable();
             $table->date('class_end')->nullable();
             $table->date('class_ext')->nullable();
